requirements.txt
opencv-python==4.10.0.84
pillow==10.4.0
torch
pillow
transformers
torchvision
numpy<2
